move items into a memoized wrapper
